Add files via upload
This project is a  responsive To-Do list using HTML, CSS, and JavaScript. The design ensures a seamless user experience with options to add, delete, and edit tasks. The list distinguishes between completed and pending tasks for better organization. Additionally, a convenient "Clear All" button is available to efficiently manage the tasks. Overall, it's a user-friendly and interactive solution for task management.